requirements.txt
ipykernel
selenium
chromedriver_autoinstaller
langchain
langchain_ollama
langchain_chroma
streamlit
python_dotenv
langchain_community
langserve
fastapi
uvicorn
sse_starlette
chromadb
faiss-cpu
nomic[local]
wikipedia
webdriver_manager
openai-whisper
torch
git+https://github.com/openai/whisper.git
neo4j-graphrag
pygraphviz
neo4j
langchain_groq
uvicorn
fastapi
python-multipart